pay request into cdk lib
